IGNITE-14748 Implemented ability to persist ordering of named list configuration elements.

+++ modules/configuration/src/main/java/org/apache/ignite/internal/configuration/util/ConfigurationUtil.java
@@ -18,16 +18,19 @@
 package org.apache.ignite.internal.configuration.util;
 
 import java.io.Serializable;
+import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Deque;
 import java.util.HashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Objects;
 import java.util.RandomAccess;
+import java.util.Set;
 import java.util.stream.Collectors;
+import org.apache.ignite.configuration.NamedListView;
 import org.apache.ignite.configuration.RootKey;
 import org.apache.ignite.internal.configuration.SuperRoot;
 import org.apache.ignite.internal.configuration.tree.ConfigurationSource;
@@ -252,14 +255,24 @@ public class ConfigurationUtil {
 
             /** {@inheritDoc} */
             @Override public void descend(ConstructableTreeNode node) {
+                boolean nodeIsNamedList = node instanceof NamedListNode;
+
+                List<String> orderedKeys = nodeIsNamedList
+                    ? new ArrayList<>(((NamedListView<?>)node).namedListKeys())
+                    : null;
+
                 for (Map.Entry<String, ?> entry : map.entrySet()) {
                     String key = entry.getKey();
                     Object val = entry.getValue();
 
                     assert val == null || val instanceof Map || val instanceof Serializable;
 
+                    // Ordering indexes must be skipped here bacause they make no sense in this context.
+                    if (key.equals(NamedListNode.ORDER_IDX))
+                        continue;
+
                     if (val == null) {
-                        if (node instanceof NamedListNode) {
+                        if (nodeIsNamedList) {
                             // Given that this particular method is applied to modify existing trees rather than
                             // creating new trees, a "hack" is required in this place. "construct" is designed to create
                             // "change" objects, thus it would just nullify named list element instead of deleting it.
@@ -268,13 +281,49 @@ public class ConfigurationUtil {
                         else
                             node.construct(key, null);
                     }
-                    else if (val instanceof Map)
+                    else if (val instanceof Map) {
+                        if (nodeIsNamedList) {
+                            // For every named list entry modification we must take its index into account.
+                            // We do this by modifying "orderedKeys" when index is explicitly passed.
+                            Object idxObj = ((Map<?, ?>)val).get(NamedListNode.ORDER_IDX);
+
+                            if (idxObj != null) {
+                                assert idxObj instanceof Integer : val;
+
+                                int idx = (Integer)idxObj;
+
+                                if (idx >= orderedKeys.size()) {
+                                    // Updates can come in arbitrary order. This means that arrays may be too small
+                                    // during batch creation. In this case we have to insert enough nulls before
+                                    // invoking "add" method for actual key.
+                                    ((ArrayList<?>)orderedKeys).ensureCapacity(idx + 1);
+
+                                    while (idx != orderedKeys.size())
+                                        orderedKeys.add(null);
+
+                                    orderedKeys.add(idx, key);
+                                }
+                                else
+                                    orderedKeys.set(idx, key);
+                            }
+                        }
+
                         node.construct(key, new InnerConfigurationSource((Map<String, ?>)val));
+                    }
                     else {
                         assert val instanceof Serializable;
 
                         node.construct(key, new LeafConfigurationSource((Serializable)val));
                     }
+                }
+
+                if (nodeIsNamedList) {
+                    NamedListNode<?> namedListNode = (NamedListNode<?>)node;
+
+                    if (orderedKeys.size() > namedListNode.size())
+                        orderedKeys = orderedKeys.subList(0, namedListNode.size());
+
+                    namedListNode.reorderKeys(orderedKeys);
                 }
             }
         }
@@ -291,87 +340,153 @@ public class ConfigurationUtil {
      * @param updates Tree with updates.
      * @return Map of changes.
      */
-    public static Map<String, Serializable> nodeToFlatMap(
-        SuperRoot curRoot,
-        SuperRoot updates
-    ) {
+    public static Map<String, Serializable> createFlattenedUpdatesMap(SuperRoot curRoot, SuperRoot updates) {
+        // Resulting map.
         Map<String, Serializable> values = new HashMap<>();
 
+        // Current methods traverses two trees at the same time. In order to reuse visitor object it's decided to
+        // use explicit stack for the left tree. We need to reuse visitor object because it accumulates full keys.
+        Deque<InnerNode> oldInnerNodesStack = new ArrayDeque<>(Set.of(curRoot));
+
         updates.traverseChildren(new KeysTrackingConfigurationVisitor<>() {
-            /** Write nulls instead of actual values. Makes sense for deletions from named lists. */
-            private boolean writeNulls;
+            /** Flag indicates that "old" and "new" trees are literally the same at the moment. */
+            private boolean singleTreeTraversal;
+
+            /**
+             * Makes sense only if {@link #singleTreeTraversal} is {@code true}. Helps distinguishing creation from
+             * deletion. Always {@code false} if {@link #singleTreeTraversal} is {@code false}.
+             */
+            private boolean deletion;
 
             /** {@inheritDoc} */
-            @Override public Map<String, Serializable> doVisitLeafNode(String key, Serializable val) {
-                if (val != null)
-                    values.put(currentKey(), writeNulls ? null : val);
+            @Override public Void doVisitLeafNode(String key, Serializable newVal) {
+                // Read same value from old tree.
+                Serializable oldVal = oldInnerNodesStack.peek().traverseChild(key, leafNodeVisitor());
 
-                return values;
+                // Do not put duplicates into the resulting map.
+                if (singleTreeTraversal || !Objects.deepEquals(oldVal, newVal))
+                    values.put(currentKey(), deletion ? null : newVal);
+
+                return null;
             }
 
             /** {@inheritDoc} */
-            @Override public Map<String, Serializable> doVisitInnerNode(String key, InnerNode node) {
-                if (node == null)
+            @Override public Void doVisitInnerNode(String key, InnerNode newNode) {
+                // Read same node from old tree.
+                InnerNode oldNode = oldInnerNodesStack.peek().traverseChild(key, innerNodeVisitor());
+
+                // Skip subtree that has not changed.
+                if (oldNode == newNode && !singleTreeTraversal)
                     return null;
 
-                node.traverseChildren(this);
+                if (oldNode == null)
+                    visitAsymmetricInnerNode(newNode, false);
+                else {
+                    oldInnerNodesStack.push(oldNode);
 
-                return values;
+                    newNode.traverseChildren(this);
+
+                    oldInnerNodesStack.pop();
+                }
+
+                return null;
             }
 
             /** {@inheritDoc} */
-            @Override public <N extends InnerNode> Map<String, Serializable> doVisitNamedListNode(String key, NamedListNode<N> node) {
-                for (String namedListKey : node.namedListKeys()) {
-                    N namedElement = node.get(namedListKey);
+            @Override public <N extends InnerNode> Void doVisitNamedListNode(String key, NamedListNode<N> newNode) {
+                // Read same named list node from old tree.
+                NamedListNode<?> oldNode = oldInnerNodesStack.peek().traverseChild(key, namedListNodeVisitor());
 
+                // Skip subtree that has not changed.
+                if (oldNode == newNode && !singleTreeTraversal)
+                    return null;
+
+                // Old keys ordering can be ignored if we either create or delete everything.
+                Map<String, Integer> oldKeysToOrderIdxMap = singleTreeTraversal ? null
+                    : keysToOrderIdx(oldNode.namedListKeys());
+
+                // New keys ordering can be ignored if we delete everything.
+                Map<String, Integer> newKeysToOrderIdxMap = deletion ? null
+                    : keysToOrderIdx(newNode.namedListKeys());
+
+                for (String namedListKey : newNode.namedListKeys()) {
                     withTracking(namedListKey, true, false, () -> {
-                        if (namedElement == null)
-                            visitDeletedNamedListElement();
-                        else
-                            namedElement.traverseChildren(this);
+                        InnerNode oldNamedElement = oldNode.get(namedListKey);
+                        InnerNode newNamedElement = newNode.get(namedListKey);
+
+                        // Deletion of nonexistent element.
+                        if (oldNamedElement == null && newNamedElement == null)
+                            return null;
+
+                        // Skip element that has not changed.
+                        // Its index can be different though, so we don't "continue" straight away.
+                        if (singleTreeTraversal || oldNamedElement != newNamedElement) {
+                            if (newNamedElement == null)
+                                visitAsymmetricInnerNode(oldNamedElement, true);
+                            else if (oldNamedElement == null)
+                                visitAsymmetricInnerNode(newNamedElement, false);
+                            else {
+                                oldInnerNodesStack.push(oldNamedElement);
+
+                                newNamedElement.traverseChildren(this);
+
+                                oldInnerNodesStack.pop();
+                            }
+                        }
+
+                        Integer newIdx = newKeysToOrderIdxMap == null ? null : newKeysToOrderIdxMap.get(namedListKey);
+                        Integer oldIdx = oldKeysToOrderIdxMap == null ? null : oldKeysToOrderIdxMap.get(namedListKey);
+
+                        // We should "persist" changed indexes only.
+                        if (newIdx != oldIdx || singleTreeTraversal || newNamedElement == null) {
+                            String orderKey = currentKey() + NamedListNode.ORDER_IDX;
+
+                            values.put(orderKey, deletion || newNamedElement == null ? null : newIdx);
+                        }
 
                         return null;
                     });
                 }
 
-                return values;
+                return null;
             }
 
             /**
-             * Here we must list all joined keys belonging to deleted element. The only way to do it is to traverse
-             * cached configuration tree and write {@code null} into all values.
+             * Here we must list all joined keys belonging to deleted or created element. The only way to do it is to
+             * traverse whole configuration tree unconditionally.
              */
-            private void visitDeletedNamedListElement() {
-                // It must be impossible to delete something inside of the element that we're currently deleting.
-                assert !writeNulls;
+            private void visitAsymmetricInnerNode(InnerNode node, boolean delete) {
+                assert !singleTreeTraversal;
+                assert node != null;
 
-                Object originalNamedElement = null;
+                oldInnerNodesStack.push(node);
+                singleTreeTraversal = true;
+                deletion = delete;
 
-                List<String> currentPath = currentPath();
+                node.traverseChildren(this);
 
-                try {
-                    // This code can in fact be better optimized for deletion scenario,
-                    // but there's no point in doing that, since the operation is so rare and it will
-                    // complicate code even more.
-                    originalNamedElement = find(currentPath, curRoot);
-                }
-                catch (KeyNotFoundException ignore) {
-                    // May happen, not a big deal. This means that element never existed in the first place.
-                }
-
-                if (originalNamedElement != null) {
-                    assert originalNamedElement instanceof InnerNode : currentPath;
-
-                    writeNulls = true;
-
-                    ((InnerNode)originalNamedElement).traverseChildren(this);
-
-                    writeNulls = false;
-                }
+                deletion = false;
+                singleTreeTraversal = false;
+                oldInnerNodesStack.pop();
             }
         });
 
         return values;
+    }
+
+    /**
+     * @param keys Ordered set of keys.
+     * @return Map that contains same keys and their positions as values.
+     */
+    private static Map<String, Integer> keysToOrderIdx(Set<String> keys) {
+        Map<String, Integer> res = new HashMap<>();
+
+        int idx = 0;
+
+        for (String key : keys)
+            res.put(key, idx++);
+
+        return res;
     }
 
     /**
@@ -539,65 +654,6 @@ public class ConfigurationUtil {
         };
     }
 
-    /**
-     * Nullifies leaves in {@code newNode} node that are equal to the corresponding leaf values in {@code curNode}.
-     * In this context we view {@code curNode} as a full configuration node with all the data, while {@code newNode}
-     * contains only updates that we plan to apply to the {@code curNode} in the future.
-     * @param curNode Node to look for definitive values.
-     * @param newNode Node to nullify matching values.
-     */
-    public static void cleanupMatchingValues(InnerNode curNode, InnerNode newNode) {
-        if (curNode == null || newNode == null)
-            return;
-
-        assert curNode.getClass() == newNode.getClass();
-
-        newNode.traverseChildren(new ConfigurationVisitor<Void>() {
-            /** {@inheritDoc} */
-            @Override public Void visitLeafNode(String key, Serializable newVal) {
-                if (Objects.deepEquals(curNode.traverseChild(key, leafNodeVisitor()), newVal))
-                    newNode.construct(key, null);
-
-                return null;
-            }
-
-            /** {@inheritDoc} */
-            @Override public Void visitInnerNode(String key, InnerNode newInnerNode) {
-                InnerNode oldInnerNode = curNode.traverseChild(key, innerNodeVisitor());
-
-                if (oldInnerNode == newInnerNode)
-                    newNode.construct(key, null);
-                else
-                    cleanupMatchingValues(oldInnerNode, newInnerNode);
-
-                return null;
-            }
-
-            /** {@inheritDoc} */
-            @Override public <N extends InnerNode> Void visitNamedListNode(String key, NamedListNode<N> newNamedList) {
-                NamedListNode<?> curNamedList = curNode.traverseChild(key, namedListNodeVisitor());
-
-                if (curNamedList == newNamedList) {
-                    newNode.construct(key, null);
-
-                    return null;
-                }
-
-                for (String name : new LinkedHashSet<>(newNamedList.namedListKeys())) {
-                    InnerNode oldElement = curNamedList.get(name);
-                    N newElement = newNamedList.get(name);
-
-                    if (oldElement == newElement)
-                        newNamedList.forceDelete(name);
-                    else
-                        cleanupMatchingValues(oldElement, newElement);
-                }
-
-                return null;
-            }
-        });
-    }
-
     /** @see #patch(ConstructableTreeNode, TraversableTreeNode) */
     private static class PatchLeafConfigurationSource implements ConfigurationSource {
         /** */
@@ -643,6 +699,9 @@ public class ConfigurationUtil {
         /** {@inheritDoc} */
         @Override public void descend(ConstructableTreeNode dstNode) {
             assert srcNode.getClass() == dstNode.getClass() : srcNode.getClass() + " : " + dstNode.getClass();
+
+            if (srcNode == dstNode)
+                return;
 
             srcNode.traverseChildren(new ConfigurationVisitor<>() {
                 @Override public Void visitLeafNode(String key, Serializable val) {
@@ -690,12 +749,15 @@ public class ConfigurationUtil {
         @Override public void descend(ConstructableTreeNode dstNode) {
             assert srcNode.getClass() == dstNode.getClass();
 
+            if (srcNode == dstNode)
+                return;
+
             for (String key : srcNode.namedListKeys()) {
                 InnerNode node = srcNode.get(key);
 
                 if (node == null)
                     ((NamedListNode<?>)dstNode).forceDelete(key); // Same as in fillFromPrefixMap.
-                else
+                else if (((NamedListView<?>)dstNode).get(key) != node)
                     dstNode.construct(key, new PatchInnerConfigurationSource(node));
             }
         }
